% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skm_umm_stations.R
\name{skm_umm_stations}
\alias{skm_umm_stations}
\title{This function prints the power stations in the Nordpool area.}
\usage{
skm_umm_stations()
}
\description{
This function is manually updated so if there are any stations,
missing then contact the maintainer.
}

