% Generated by roxygen2 (4.0.1): do not edit by hand
\name{skm_ummquery}
\alias{skm_ummquery}
\title{UMM Query the SKM Syspower API.}
\usage{
skm_ummquery(user_id, user_password, interval, start_time, end_time, type,
  areas, internalorfuels, headers = "yes", accrow = "no",
  data_format = "se")
}
\arguments{
\item{user_id}{character value. Login name to SKM syspower.}

\item{user_password}{Password to SKM syspower.}

\item{interval}{Character value. "hour", "day" or "week". Weekly interval will be parsed as ISO week (see package ISOweek for parsing this data_format).}

\item{start_time}{Character value with format YYYY-mm-dd, dd-mm-YYYY, dd.mm.YYYY or mm/dd/YYYY.}

\item{end_time}{Character value with format YYYY-mm-dd, dd-mm-YYYY, dd.mm.YYYY or mm/dd/YYYY (YYYY-mm-dd hh:mm). "0" (wich is the default) will return today's date or the latest non-empty date.}

\item{type}{Character value. production, transmission or station.}

\item{areas}{Character string or vector. If type == production then Nordpool, DK1, DK2, SE1, SE2, SE3, SE4, NO1, NO2, NO3, NO4, NO5, EE, LV, LT, FI are valid. If type == transmission then Belarus, Estonia, Denmark, Finland, Germany, Latvia, Lithuania, Netherlands, Norway, Poland, Russia, Sweden are valid.}

\item{internalorfuels}{Character vector. If type == production then Nuclear, Wind, Coal, Gas, Hydro, Oil, Other, ReducedLoad. If type == transmission then "yes" or "no" for include internal lines.}

\item{accrow}{character value. "yes" or "no" for include accumulated row.}

\item{data_format}{"no2" (default). (Later:"no", "dk", "se", "us" might be added).}

\item{headers}{"yes" (default)}
}
\description{
This function creates a web query url for SKM umms.
}

