% Generated by roxygen2 (4.0.1): do not edit by hand
\name{skm_umm_stations}
\alias{skm_umm_stations}
\title{This function prints the power stations in the Nordpool area.}
\usage{
skm_umm_stations()
}
\description{
This function is manually updated so if there are any stations,
missing then contact the maintainer.
}

