% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skm_webquery.R
\name{skm_webquery}
\alias{skm_webquery}
\title{Query SKM Syspower API.}
\usage{
skm_webquery(user_id, user_password, series_name, interval, start_time,
  end_time = "0", empty_data = "no", currency = NULL)
}
\arguments{
\item{user_id}{character value. Login name to SKM syspower.}

\item{user_password}{Password to SKM syspower.}

\item{series_name}{character value. Ex: "NPENOYR15", "SPOT", "EEX", "EURSEKCAL15", "EURDKKQ414" or another value from the webquery on syspower.skm.no}

\item{interval}{Character value. "hour", "day" or "week". Weekly interval will be parsed as ISO week (see package ISOweek for parsing this data_format).}

\item{start_time}{Character value with format YYYY-mm-dd, dd-mm-YYYY, dd.mm.YYYY or mm/dd/YYYY.}

\item{end_time}{Character value with format YYYY-mm-dd, dd-mm-YYYY, dd.mm.YYYY or mm/dd/YYYY. "0" (wich is the default) will return today's date or the latest non-empty date.}

\item{empty_data}{"yes" or "no" (later: "remove" and "replace" will be added.)}

\item{currency}{NULL, "EUR", "DKK", "GPL", "SEK", "USD" or "NOK".}
}
\description{
This function creates a web query url and reads financial and
other data from SKM syspower into a data.frame.
}

